requirements-deploy.txt
# LegalEase AI Advanced - Production Requirements
# Core Flask Requirements
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Session==0.5.0
Werkzeug==2.3.7

# AI & ML Requirements
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Production Server
gunicorn==21.2.0

# Optional Features (uncomment if needed)
# flask-socketio==5.3.0
# flask-cors==4.0.0
# pdfplumber==0.9.0
# python-docx==1.1.0
# Pillow==10.0.0